Redux of the concept for use as directory tree.
Going to use this to build a directory tree of keys for Addendum. Need
to be able to distinguish between has and exists. Well, not really, but
that came up as I worked on it.

Removed the array of mapped items and made it just a single object. If
the user wants an array, then put an array in the map.

Added recursive matching to globbing. Renamed match to glob and it
returns a list of keys. You can then use the other methods to retreive
values even using `Array.filter` and `Array.map` to make them single
statement transforms.

Converted to ES6.

Closes #11.
Closes #10.
Closes #9.